consolidate all CSS to CSS folder.
